move into auth namespace
