+++ packages/cli/src/platform/auth/index_.ts
@@ -1,2 +1,4 @@
 export * from './$'
+export * from './login'
+export * from './logout'
 export * from './show'
