+++ packages/cli/src/platform/index_.ts
@@ -2,7 +2,7 @@ export * from './$'
 export * from './accelerate'
 export * from './apikey'
 export * from './auth'
-export * from './login'
-export * from './logout'
+export * from './auth/login'
+export * from './auth/logout'
 export * from './project'
 export * from './workspace'
